application/facematcher_deploy/requirements_extended.txt
altair==5.1.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dlib==19.24.2
extra-streamlit-components==0.1.60
face-recognition==1.3.0
face-recognition-models==0.3.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.2
google-auth-oauthlib==1.1.0
gspread==5.9.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.4
psutil==5.9.5
psycopg2-binary==2.9.9
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.4
rsa==4.9
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.20
streamlit==1.27.2
streamlit-authenticator==0.2.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
validators==0.22.0
zipp==3.17.0
